Move back backend/config & backend/core to backend pkg
