+++ backend/backend_test.go
@@ -1,6 +1,6 @@
 // +build integration,race
 
-package core
+package backend
 
 import (
 	"encoding/base64"
@@ -11,7 +11,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/sensu/sensu-go/backend/config"
 	"github.com/sensu/sensu-go/backend/etcd"
 	"github.com/sensu/sensu-go/testing/testutil"
 	"github.com/sensu/sensu-go/transport"
@@ -46,9 +45,9 @@ func TestBackendHTTPListener(t *testing.T) {
 	}{
 		{"HTTP", "http", "ws", nil},
 		{"HTTPS", "https", "wss", &types.TLSOptions{
-			CertFile:           "../../util/ssl/etcd1.pem",
-			KeyFile:            "../../util/ssl/etcd1-key.pem",
-			TrustedCAFile:      "../../util/ssl/ca.pem",
+			CertFile:           "../util/ssl/etcd1.pem",
+			KeyFile:            "../util/ssl/etcd1-key.pem",
+			TrustedCAFile:      "../util/ssl/ca.pem",
 			InsecureSkipVerify: false}},
 	}
 	// tc = Test Case
@@ -73,7 +72,7 @@ func TestBackendHTTPListener(t *testing.T) {
 
 			b := &Backend{}
 
-			err = b.NewBackend(&config.Backend{
+			err = b.LoadConfig(&Config{
 				AgentHost:                   "127.0.0.1",
 				AgentPort:                   agentPort,
 				APIHost:                     "127.0.0.1",
