+++ build.ps1
@@ -94,12 +94,13 @@ function build_binary([string]$goos, [string]$goarch, [string]$bin, [string]$cmd
         $ldflags = $ldflags + " -X $version_pkg.PreReleaseIdentifier=$prerelease"
     }
 
-    $main_pkg = "cmd/..."
+    # sensu-agent & sensuctl main packages are still in a subdirectoy cmd package
+    $main_pkg = "$bin/cmd/..."
     If ($bin -eq "backend") {
-        $main_pkg = ""
+        $main_pkg = "cmd/$cmd_name"
     }
 
-    go build -ldflags "$ldflags" -o $outfile "$REPO_PATH/$bin/$main_pkg"
+    go build -ldflags "$ldflags" -o $outfile "$REPO_PATH/$main_pkg"
     If ($LASTEXITCODE -ne 0) {
         echo "Failed to build $outfile..."
         exit 1
