+++ cmd/sensu-backend/main.go
@@ -3,11 +3,11 @@ package main
 import (
 	_ "net/http/pprof"
 
+	"github.com/sensu/sensu-go/backend"
 	"github.com/sensu/sensu-go/backend/cmd"
-	"github.com/sensu/sensu-go/backend/core"
 )
 
 func main() {
-	backend := &core.Backend{}
+	backend := &backend.Backend{}
 	cmd.Execute(backend)
 }
